Issue #1202: Change gpi_set_signal_value_long(long) to gpi_set_signal_value_int(int32_t)
- vpi_put_value() is limited to 32-bit, so use int32_t instead of long.
- Fix 32-bit platform compilation warning: comparison of integer expressions of different signedness.

+++ cocotb/share/lib/gpi/gpi_priv.h
@@ -164,7 +164,7 @@ public:
 
     int m_length;
 
-    virtual int set_signal_value(const long value, gpi_set_action_t action) = 0;
+    virtual int set_signal_value(const int32_t value, gpi_set_action_t action) = 0;
     virtual int set_signal_value(const double value, gpi_set_action_t action) = 0;
     virtual int set_signal_value_str(std::string &value, gpi_set_action_t action) = 0;
     virtual int set_signal_value_binstr(std::string &value, gpi_set_action_t action) = 0;
